Rename <PrerenderedDropdown /> to <Dropdown />
